Start 02/07/2021 | 21w05a2 | 0.7.23

+++ TRAP/src/Graphics/Buffers/VertexBuffer.cpp
@@ -74,7 +74,7 @@ void TRAP::Graphics::VertexBuffer::SetLayout(const BufferLayout& layout)
 
 //-------------------------------------------------------------------------------------------------------------------//
 
-uint32_t TRAP::Graphics::VertexBuffer::GetSize() const
+uint64_t TRAP::Graphics::VertexBuffer::GetSize() const
 {
 	return m_vertexBuffer->GetSize();
 }
